+++ source/platforms/github/GitHubGit.ts
@@ -106,25 +106,36 @@ export default async function gitDSLForGitHub(api: GitHubAPI): Promise<GitDSL> {
       const backAStepPath = pathSteps.length <= 2 ? path : pathSteps.slice(0, pathSteps.length - 1).join("/")
 
       const diff: any = {
-        after: jsonpointer.get(after, backAStepPath),
-        before: jsonpointer.get(before, backAStepPath),
+        after: jsonpointer.get(after, backAStepPath) || null,
+        before: jsonpointer.get(before, backAStepPath) || null,
       }
+
+      const counterpart = (other) => {
+        if (Array.isArray(other)) {
+          return []
+        } else if (isobject(diff.after)) {
+          return {}
+        }
+        return null
+      }
+
+      const beforeValue = diff.before || counterpart(diff.after)
+      const afterValue = diff.after || counterpart(diff.before)
 
       // If they both are arrays, add some extra metadata about what was
       // added or removed. This makes it really easy to act on specific
       // changes to JSON DSLs
 
-      if (Array.isArray(diff.after) && Array.isArray(diff.before)) {
-        const arrayBefore = diff.before as any[]
-        const arrayAfter = diff.after as any[]
+      if (Array.isArray(afterValue) && Array.isArray(beforeValue)) {
+        const arrayBefore = beforeValue as any[]
+        const arrayAfter = afterValue as any[]
 
         diff.added = arrayAfter.filter(o => !includes(arrayBefore, o))
         diff.removed = arrayBefore.filter(o => !includes(arrayAfter, o))
-
         // Do the same, but for keys inside an object if they both are objects.
-      } else if (isobject(diff.after) && isobject(diff.before)) {
-        const beforeKeys = keys(diff.before) as string[]
-        const afterKeys = keys(diff.after) as string[]
+      } else if (isobject(afterValue) && isobject(beforeValue)) {
+        const beforeKeys = keys(beforeValue) as string[]
+        const afterKeys = keys(afterValue) as string[]
         diff.added = afterKeys.filter(o => !includes(beforeKeys, o))
         diff.removed = beforeKeys.filter(o => !includes(afterKeys, o))
       }