+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### Master
 
+### 2.0.0-alpha.15
+
 - Updates `diffForFile`, `JSONPatchForFile`, and `JSONDiffForFile` to include created and removed files - #368 - bdotdub
 
 ### 2.0.0-alpha.14